+++ R/as_mat_luv.R
@@ -4,7 +4,8 @@
 #' it seems reasonable to compile these coordinates using a
 #' data frame.
 #'
-#' @param df_hcl `data.frame` with variables `h`, `c`, `l`
+#' @param df_hcl     `data.frame` with variables `h`, `c`, `l`.
+#' @param chroma_min `numeric` minimum value for chroma, helps preserve hue.
 #'
 #' @return `matrix` with columns, `l`, `u`, `v` and same number
 #'   of rows as `df_hcl`
@@ -13,7 +14,10 @@
 #'   as_mat_luv(df_hcl_blues)
 #' @export
 #'
-as_mat_luv <- function(df_hcl) {
+as_mat_luv <- function(df_hcl, chroma_min = 1.e-2) {
+
+  # set chroma floor
+  df_hcl[["c"]] <- pmax(df_hcl[["c"]], chroma_min)
 
   hcl_mat <- as.matrix(df_hcl[, c("h", "c", "l")])